Run newman via external command
Since the newman js API doesn't work very well after repeat invocations
due to all of the singletons, not to mention the memory leak, instead
use a child_process to run newman as a command.

Also tidied up the logging, colours and add a "-r" option for number of
repetitions (not to be confused with Postman collection iterations).